Remove jsas and jswbxml from apps/email/js/ext (they're included in gaia-email-opt now)
