+++ apps/email/index.html
@@ -43,16 +43,8 @@
   <script type="text/javascript" src="js/compose-cards.js"></script>
   <script type="text/javascript" src="js/setup-cards.js"></script>
   <!-- This one is an optimized vendor library containing our mail backend. -->
-  <script type="text/javascript" src="js/ext/gaia-email-opt.js"></script>
-  <!-- These are some temporary files for ActiveSync that will eventually go
-       into gaia-email-opt.js. -->
   <script type="application/javascript;version=1.8"
-          src="js/ext/wbxml.js"></script>
-  <script type="application/javascript;version=1.8"
-          src="js/ext/activesync/codepages.js"></script>
-  <script type="application/javascript;version=1.8"
-          src="js/ext/activesync/protocol.js"></script>
-
+          src="js/ext/gaia-email-opt.js"></script>
   <script type="text/javascript" src="js/mail-app.js"></script>
 </head>
