--- apps/email/js/ext/activesync/codepages.js
@@ -1,685 +1,0 @@
-/* Copyright 2012 Mozilla Foundation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-(function (root, factory) {
-  if (typeof exports === "object")
-    module.exports = factory(require("wbxml"));
-  else if (typeof define === "function" && define.amd)
-    define(["wbxml"], factory);
-  else
-    root.ActiveSyncCodepages = factory(WBXML);
-}(this, function(WBXML) {
-  let codepages = {
-    AirSync: {
-      Tags: {
-        Sync:              0x0005,
-        Responses:         0x0006,
-        Add:               0x0007,
-        Change:            0x0008,
-        Delete:            0x0009,
-        Fetch:             0x000A,
-        SyncKey:           0x000B,
-        ClientId:          0x000C,
-        ServerId:          0x000D,
-        Status:            0x000E,
-        Collection:        0x000F,
-        Class:             0x0010,
-        CollectionId:      0x0012,
-        GetChanges:        0x0013,
-        MoreAvailable:     0x0014,
-        WindowSize:        0x0015,
-        Commands:          0x0016,
-        Options:           0x0017,
-        FilterType:        0x0018,
-        Conflict:          0x001B,
-        Collections:       0x001C,
-        ApplicationData:   0x001D,
-        DeletesAsMoves:    0x001E,
-        Supported:         0x0020,
-        SoftDelete:        0x0021,
-        MIMESupport:       0x0022,
-        MIMETruncation:    0x0023,
-        Wait:              0x0024,
-        Limit:             0x0025,
-        Partial:           0x0026,
-        ConversationMode:  0x0027,
-        MaxItems:          0x0028,
-        HeartbeatInterval: 0x0029,
-      },
-    },
-
-    Contacts: {
-      Tags: {
-        Anniversary:               0x0105,
-        AssistantName:             0x0106,
-        AssistantPhoneNumber:      0x0107,
-        Birthday:                  0x0108,
-        Business2PhoneNumber:      0x010C,
-        BusinessAddressCity:       0x010D,
-        BusinessAddressCountry:    0x010E,
-        BusinessAddressPostalCode: 0x010F,
-        BusinessAddressState:      0x0110,
-        BusinessAddressStreet:     0x0111,
-        BusinessFaxNumber:         0x0112,
-        BusinessPhoneNumber:       0x0113,
-        CarPhoneNumber:            0x0114,
-        Categories:                0x0115,
-        Category:                  0x0116,
-        Children:                  0x0117,
-        Child:                     0x0118,
-        CompanyName:               0x0119,
-        Department:                0x011A,
-        Email1Address:             0x011B,
-        Email2Address:             0x011C,
-        Email3Address:             0x011D,
-        FileAs:                    0x011E,
-        FirstName:                 0x011F,
-        Home2PhoneNumber:          0x0120,
-        HomeAddressCity:           0x0121,
-        HomeAddressCountry:        0x0122,
-        HomeAddressPostalCode:     0x0123,
-        HomeAddressState:          0x0124,
-        HomeAddressStreet:         0x0125,
-        HomeFaxNumber:             0x0126,
-        HomePhoneNumber:           0x0127,
-        JobTitle:                  0x0128,
-        LastName:                  0x0129,
-        MiddleName:                0x012A,
-        MobilePhoneNumber:         0x012B,
-        OfficeLocation:            0x012C,
-        OtherAddressCity:          0x012D,
-        OtherAddressCountry:       0x012E,
-        OtherAddressPostalCode:    0x012F,
-        OtherAddressState:         0x0130,
-        OtherAddressStreet:        0x0131,
-        PagerNumber:               0x0132,
-        RadioPhoneNumber:          0x0133,
-        Spouse:                    0x0134,
-        Suffix:                    0x0135,
-        Title:                     0x0136,
-        WebPage:                   0x0137,
-        YomiCompanyName:           0x0138,
-        YomiFirstName:             0x0139,
-        YomiLastName:              0x013A,
-        Picture:                   0x013C,
-        Alias:                     0x013D,
-        WeightedRank:              0x013E,
-      },
-    },
-
-    Email: {
-      Tags: {
-        DateReceived:            0x020F,
-        DisplayTo:               0x0211,
-        Importance:              0x0212,
-        MessageClass:            0x0213,
-        Subject:                 0x0214,
-        Read:                    0x0215,
-        To:                      0x0216,
-        Cc:                      0x0217,
-        From:                    0x0218,
-        ReplyTo:                 0x0219,
-        AllDayEvent:             0x021A,
-        Categories:              0x021B,
-        Category:                0x021C,
-        DTStamp:                 0x021D,
-        EndTime:                 0x021E,
-        InstanceType:            0x021F,
-        BusyStatus:              0x0220,
-        Location:                0x0221,
-        MeetingRequest:          0x0222,
-        Organizer:               0x0223,
-        RecurrenceId:            0x0224,
-        Reminder:                0x0225,
-        ResponseRequested:       0x0226,
-        Recurrences:             0x0227,
-        Recurrence:              0x0228,
-        Recurrence_Type:         0x0229,
-        Recurrence_Until:        0x022A,
-        Recurrence_Occurrences:  0x022B,
-        Recurrence_Interval:     0x022C,
-        Recurrence_DayOfWeek:    0x022D,
-        Recurrence_DayOfMonth:   0x022E,
-        Recurrence_WeekOfMonth:  0x022F,
-        Recurrence_MonthOfYear:  0x0230,
-        StartTime:               0x0231,
-        Sensitivity:             0x0232,
-        TimeZone:                0x0233,
-        GlobalObjId:             0x0234,
-        ThreadTopic:             0x0235,
-        InternetCPID:            0x0239,
-        Flag:                    0x023A,
-        Status:                  0x023B,
-        ContentClass:            0x023C,
-        FlagType:                0x023D,
-        CompleteTime:            0x023E,
-        DisallowNewTimeProposal: 0x023F,
-      },
-    },
-
-    Calendar: {
-      Tags: {
-        TimeZone:                  0x0405,
-        AllDayEvent:               0x0406,
-        Attendees:                 0x0407,
-        Attendee:                  0x0408,
-        Email:                     0x0409,
-        Name:                      0x040A,
-        BusyStatus:                0x040D,
-        Categories:                0x040E,
-        Category:                  0x040F,
-        DtStamp:                   0x0411,
-        EndTime:                   0x0412,
-        Exception:                 0x0413,
-        Exceptions:                0x0414,
-        Deleted:                   0x0415,
-        ExceptionStartTime:        0x0416,
-        Location:                  0x0417,
-        MeetingStatus:             0x0418,
-        OrganizerEmail:            0x0419,
-        OrganizerName:             0x041A,
-        Recurrence:                0x041B,
-        Type:                      0x041C,
-        Until:                     0x041D,
-        Occurrences:               0x041E,
-        Interval:                  0x041F,
-        DayOfWeek:                 0x0420,
-        DayOfMonth:                0x0421,
-        WeekOfMonth:               0x0422,
-        MonthOfYear:               0x0423,
-        Reminder:                  0x0424,
-        Sensitivity:               0x0425,
-        Subject:                   0x0426,
-        StartTime:                 0x0427,
-        UID:                       0x0428,
-        AttendeeStatus:            0x0429,
-        AttendeeType:              0x042A,
-        DisallowNewTimeProposal:   0x0433,
-        ResponseRequested:         0x0434,
-        AppointmentReplyTime:      0x0435,
-        ResponseType:              0x0436,
-        CalendarType:              0x0437,
-        IsLeapMonth:               0x0438,
-        FirstDayOfWeek:            0x0439,
-        OnlineMeetingConfLink:     0x043A,
-        OnlineMeetingExternalLink: 0x043B,
-      },
-    },
-
-    Move: {
-      Tags: {
-        MoveItems: 0x0505,
-        Move:      0x0506,
-        SrcMsgId:  0x0507,
-        SrcFldId:  0x0508,
-        DstFldId:  0x0509,
-        Response:  0x050A,
-        Status:    0x050B,
-        DstMsgId:  0x050C,
-      },
-    },
-
-    ItemEstimate: {
-      Tags: {
-        GetItemEstimate: 0x0605,
-        Version:         0x0606,
-        Collections:     0x0607,
-        Collection:      0x0608,
-        Class:           0x0609,
-        CollectionId:    0x060A,
-        DateTime:        0x060B,
-        Estimate:        0x060C,
-        Response:        0x060D,
-        Status:          0x060E,
-      },
-    },
-
-    FolderHierarchy: {
-      Tags: {
-        DisplayName:  0x0707,
-        ServerId:     0x0708,
-        ParentId:     0x0709,
-        Type:         0x070A,
-        Status:       0x070C,
-        Changes:      0x070E,
-        Add:          0x070F,
-        Delete:       0x0710,
-        Update:       0x0711,
-        SyncKey:      0x0712,
-        FolderCreate: 0x0713,
-        FolderDelete: 0x0714,
-        FolderUpdate: 0x0715,
-        FolderSync:   0x0716,
-        Count:        0x0717,
-      },
-    },
-
-    MeetingResponse: {
-      Tags: {
-        CalendarId:      0x0805,
-        CollectionId:    0x0806,
-        MeetingResponse: 0x0807,
-        RequestId:       0x0808,
-        Request:         0x0809,
-        Result:          0x080A,
-        Status:          0x080B,
-        UserResponse:    0x080C,
-        InstanceId:      0x080E,
-      },
-    },
-
-    Tasks: {
-      Tags: {
-        Categories:             0x0908,
-        Category:               0x0909,
-        Complete:               0x090A,
-        DateCompleted:          0x090B,
-        DueDate:                0x090C,
-        UtcDueDate:             0x090D,
-        Importance:             0x090E,
-        Recurrence:             0x090F,
-        Recurrence_Type:        0x0910,
-        Recurrence_Start:       0x0911,
-        Recurrence_Until:       0x0912,
-        Recurrence_Occurrences: 0x0913,
-        Recurrence_Interval:    0x0914,
-        Recurrence_DayOfMonth:  0x0915,
-        Recurrence_DayOfWeek:   0x0916,
-        Recurrence_WeekOfMonth: 0x0917,
-        Recurrence_MonthOfYear: 0x0918,
-        Recurrence_Regenerate:  0x0919,
-        Recurrence_DeadOccur:   0x091A,
-        ReminderSet:            0x091B,
-        ReminderTime:           0x091C,
-        Sensitivity:            0x091D,
-        StartDate:              0x091E,
-        UtcStartDate:           0x091F,
-        Subject:                0x0920,
-        OrdinalDate:            0x0922,
-        SubOrdinalDate:         0x0923,
-        CalendarType:           0x0924,
-        IsLeapMonth:            0x0925,
-        FirstDayOfWeek:         0x0926,
-      },
-    },
-
-    ResolveRecipients: {
-      Tags: {
-        ResolveRecipients:      0x0A05,
-        Response:               0x0A06,
-        Status:                 0x0A07,
-        Type:                   0x0A08,
-        Recipient:              0x0A09,
-        DisplayName:            0x0A0A,
-        EmailAddress:           0x0A0B,
-        Certificates:           0x0A0C,
-        Certificate:            0x0A0D,
-        MiniCertificate:        0x0A0E,
-        Options:                0x0A0F,
-        To:                     0x0A10,
-        CertificateRetrieval:   0x0A11,
-        RecipientCount:         0x0A12,
-        MaxCertificates:        0x0A13,
-        MaxAmbiguousRecipients: 0x0A14,
-        CertificateCount:       0x0A15,
-        Availability:           0x0A16,
-        StartTime:              0x0A17,
-        EndTime:                0x0A18,
-        MergedFreeBusy:         0x0A19,
-        Picture:                0x0A1A,
-        MaxSize:                0x0A1B,
-        Data:                   0x0A1C,
-        MaxPictures:            0x0A1D,
-      },
-    },
-
-    ValidateCert: {
-      Tags: {
-        ValidateCert:     0x0B05,
-        Certificates:     0x0B06,
-        Certificate:      0x0B07,
-        CertificateChain: 0x0B08,
-        CheckCRL:         0x0B09,
-        Status:           0x0B0A,
-      },
-    },
-
-    Contacts2: {
-      Tags: {
-        CustomerId:       0x0C05,
-        GovernmentId:     0x0C06,
-        IMAddress:        0x0C07,
-        IMAddress2:       0x0C08,
-        IMAddress3:       0x0C09,
-        ManagerName:      0x0C0A,
-        CompanyMainPhone: 0x0C0B,
-        AccountName:      0x0C0C,
-        NickName:         0x0C0D,
-        MMS:              0x0C0E,
-      },
-    },
-
-    Ping: {
-      Tags: {
-        Ping:              0x0D05,
-        AutdState:         0x0D06,
-        Status:            0x0D07,
-        HeartbeatInterval: 0x0D08,
-        Folders:           0x0D09,
-        Folder:            0x0D0A,
-        Id:                0x0D0B,
-        Class:             0x0D0C,
-        MaxFolders:        0x0D0D,
-      },
-    },
-
-    Provision: {
-      Tags: {
-        Provision:                                0x0E05,
-        Policies:                                 0x0E06,
-        Policy:                                   0x0E07,
-        PolicyType:                               0x0E08,
-        PolicyKey:                                0x0E09,
-        Data:                                     0x0E0A,
-        Status:                                   0x0E0B,
-        RemoteWipe:                               0x0E0C,
-        EASProvisionDoc:                          0x0E0D,
-        DevicePasswordEnabled:                    0x0E0E,
-        AlphanumericDevicePasswordRequired:       0x0E0F,
-        DeviceEncryptionEnabled:                  0x0E10,
-        RequireStorageCardEncryption:             0x0E10,
-        PasswordRecoveryEnabled:                  0x0E11,
-        AttachmentsEnabled:                       0x0E13,
-        MinDevicePasswordLength:                  0x0E14,
-        MaxInactivityTimeDeviceLock:              0x0E15,
-        MaxDevicePasswordFailedAttempts:          0x0E16,
-        MaxAttachmentSize:                        0x0E17,
-        AllowSimpleDevicePassword:                0x0E18,
-        DevicePasswordExpiration:                 0x0E19,
-        DevicePasswordHistory:                    0x0E1A,
-        AllowStorageCard:                         0x0E1B,
-        AllowCamera:                              0x0E1C,
-        RequireDeviceEncryption:                  0x0E1D,
-        AllowUnsignedApplications:                0x0E1E,
-        AllowUnsignedInstallationPackages:        0x0E1F,
-        MinDevicePasswordComplexCharacters:       0x0E20,
-        AllowWiFi:                                0x0E21,
-        AllowTextMessaging:                       0x0E22,
-        AllowPOPIMAPEmail:                        0x0E23,
-        AllowBluetooth:                           0x0E24,
-        AllowIrDA:                                0x0E25,
-        RequireManualSyncWhenRoaming:             0x0E26,
-        AllowDesktopSync:                         0x0E27,
-        MaxCalendarAgeFilter:                     0x0E28,
-        AllowHTMLEmail:                           0x0E29,
-        MaxEmailAgeFilter:                        0x0E2A,
-        MaxEmailBodyTruncationSize:               0x0E2B,
-        MaxEmailHTMLBodyTruncationSize:           0x0E2C,
-        RequireSignedSMIMEMessages:               0x0E2D,
-        RequireEncryptedSMIMEMessages:            0x0E2E,
-        RequireSignedSMIMEAlgorithm:              0x0E2F,
-        RequireEncryptionSMIMEAlgorithm:          0x0E30,
-        AllowSMIMEEncryptionAlgorithmNegotiation: 0x0E31,
-        AllowSMIMESoftCerts:                      0x0E32,
-        AllowBrowser:                             0x0E33,
-        AllowConsumerEmail:                       0x0E34,
-        AllowRemoteDesktop:                       0x0E35,
-        AllowInternetSharing:                     0x0E36,
-        UnapprovedInROMApplicationList:           0x0E37,
-        ApplicationName:                          0x0E38,
-        ApprovedApplicationList:                  0x0E39,
-        Hash:                                     0x0E3A,
-      },
-    },
-
-    Search: {
-      Tags: {
-        Search:         0x0F05,
-        Store:          0x0F07,
-        Name:           0x0F08,
-        Query:          0x0F09,
-        Options:        0x0F0A,
-        Range:          0x0F0B,
-        Status:         0x0F0C,
-        Response:       0x0F0D,
-        Result:         0x0F0E,
-        Properties:     0x0F0F,
-        Total:          0x0F10,
-        EqualTo:        0x0F11,
-        Value:          0x0F12,
-        And:            0x0F13,
-        Or:             0x0F14,
-        FreeText:       0x0F15,
-        DeepTraversal:  0x0F17,
-        LongId:         0x0F18,
-        RebuildResults: 0x0F19,
-        LessThan:       0x0F1A,
-        GreaterThan:    0x0F1B,
-        UserName:       0x0F1E,
-        Password:       0x0F1F,
-        ConversationId: 0x0F20,
-        Picture:        0x0F21,
-        MaxSize:        0x0F22,
-        MaxPictures:    0x0F23,
-      },
-    },
-
-    GAL: {
-      Tags: {
-        DisplayName:  0x1005,
-        Phone:        0x1006,
-        Office:       0x1007,
-        Title:        0x1008,
-        Company:      0x1009,
-        Alias:        0x100A,
-        FirstName:    0x100B,
-        LastName:     0x100C,
-        HomePhone:    0x100D,
-        MobilePhone:  0x100E,
-        EmailAddress: 0x100F,
-        Picture:      0x1010,
-        Status:       0x1011,
-        Data:         0x1012,
-      },
-    },
-
-    AirSyncBase: {
-      Tags: {
-        BodyPreference:     0x1105,
-        Type:               0x1106,
-        TruncationSize:     0x1107,
-        AllOrNone:          0x1108,
-        Body:               0x110A,
-        Data:               0x110B,
-        EstimatedDataSize:  0x110C,
-        Truncated:          0x110D,
-        Attachments:        0x110E,
-        Attachment:         0x110F,
-        DisplayName:        0x1110,
-        FileReference:      0x1111,
-        Method:             0x1112,
-        ContentId:          0x1113,
-        ContentLocation:    0x1114,
-        IsInline:           0x1115,
-        NativeBodyType:     0x1116,
-        ContentType:        0x1117,
-        Preview:            0x1118,
-        BodyPartPreference: 0x1119,
-        BodyPart:           0x111A,
-        Status:             0x111B,
-      },
-    },
-
-    Settings: {
-      Tags: {
-        Settings:                    0x1205,
-        Status:                      0x1206,
-        Get:                         0x1207,
-        Set:                         0x1208,
-        Oof:                         0x1209,
-        OofState:                    0x120A,
-        StartTime:                   0x120B,
-        EndTime:                     0x120C,
-        OofMessage:                  0x120D,
-        AppliesToInternal:           0x120E,
-        AppliesToExternalKnown:      0x120F,
-        AppliesToExternalUnknown:    0x1210,
-        Enabled:                     0x1211,
-        ReplyMessage:                0x1212,
-        BodyType:                    0x1213,
-        DevicePassword:              0x1214,
-        Password:                    0x1215,
-        DeviceInformation:           0x1216,
-        Model:                       0x1217,
-        IMEI:                        0x1218,
-        FriendlyName:                0x1219,
-        OS:                          0x121A,
-        OSLanguage:                  0x121B,
-        PhoneNumber:                 0x121C,
-        UserInformation:             0x121D,
-        EmailAddresses:              0x121E,
-        SmtpAddress:                 0x121F,
-        UserAgent:                   0x1220,
-        EnableOutboundSMS:           0x1221,
-        MobileOperator:              0x1222,
-        PrimarySmtpAddress:          0x1223,
-        Accounts:                    0x1224,
-        Account:                     0x1225,
-        AccountId:                   0x1226,
-        AccountName:                 0x1227,
-        UserDisplayName:             0x1228,
-        SendDisabled:                0x1229,
-        RightsManagementInformation: 0x122B,
-      },
-    },
-
-    DocumentLibrary: {
-      Tags: {
-        LinkId:           0x1305,
-        DisplayName:      0x1306,
-        IsFolder:         0x1307,
-        CreationDate:     0x1308,
-        LastModifiedDate: 0x1309,
-        IsHidden:         0x130A,
-        ContentLength:    0x130B,
-        ContentType:      0x130C,
-      },
-    },
-
-    ItemOperations: {
-      Tags: {
-        ItemOperations:      0x1405,
-        Fetch:               0x1406,
-        Store:               0x1407,
-        Options:             0x1408,
-        Range:               0x1409,
-        Total:               0x140A,
-        Properties:          0x140B,
-        Data:                0x140C,
-        Status:              0x140D,
-        Response:            0x140E,
-        Version:             0x140F,
-        Schema:              0x1410,
-        Part:                0x1411,
-        EmptyFolderContents: 0x1412,
-        DeleteSubFolders:    0x1413,
-        UserName:            0x1414,
-        Password:            0x1415,
-        Move:                0x1416,
-        DstFldId:            0x1417,
-        ConversationId:      0x1418,
-        MoveAlways:          0x1419,
-      },
-    },
-
-    ComposeMail: {
-      Tags: {
-        SendMail:        0x1505,
-        SmartForward:    0x1506,
-        SmartReply:      0x1507,
-        SaveInSentItems: 0x1508,
-        ReplaceMime:     0x1509,
-        Source:          0x150B,
-        FolderId:        0x150C,
-        ItemId:          0x150D,
-        LongId:          0x150E,
-        InstanceId:      0x150F,
-        Mime:            0x1510,
-        ClientId:        0x1511,
-        Status:          0x1512,
-        AccountId:       0x1513,
-      },
-    },
-
-    Email2: {
-      Tags: {
-        UmCallerID:            0x1605,
-        UmUserNotes:           0x1606,
-        UmAttDuration:         0x1607,
-        UmAttOrder:            0x1608,
-        ConversationId:        0x1609,
-        ConversationIndex:     0x160A,
-        LastVerbExecuted:      0x160B,
-        LastVerbExecutionTime: 0x160C,
-        ReceivedAsBcc:         0x160D,
-        Sender:                0x160E,
-        CalendarType:          0x160F,
-        IsLeapMonth:           0x1610,
-        AccountId:             0x1611,
-        FirstDayOfWeek:        0x1612,
-        MeetingMessageType:    0x1613,
-      },
-    },
-
-    Notes: {
-      Tags: {
-        Subject:          0x1705,
-        MessageClass:     0x1706,
-        LastModifiedDate: 0x1707,
-        Categories:       0x1708,
-        Category:         0x1709,
-      },
-    },
-
-    RightsManagement: {
-      Tags: {
-        RightsManagementSupport:            0x1805,
-        RightsManagementTemplates:          0x1806,
-        RightsManagementTemplate:           0x1807,
-        RightsManagementLicense:            0x1808,
-        EditAllowed:                        0x1809,
-        ReplyAllowed:                       0x180A,
-        ReplyAllAllowed:                    0x180B,
-        ForwardAllowed:                     0x180C,
-        ModifyRecipientsAllowed:            0x180D,
-        ExtractAllowed:                     0x180E,
-        PrintAllowed:                       0x180F,
-        ExportAllowed:                      0x1810,
-        ProgrammaticAccessAllowed:          0x1811,
-        Owner:                              0x1812,
-        ContentExpiryDate:                  0x1813,
-        TemplateID:                         0x1814,
-        TemplateName:                       0x1815,
-        TemplateDescription:                0x1816,
-        ContentOwner:                       0x1817,
-        RemoveRightsManagementDistribution: 0x1818,
-      },
-    },
-  };
-
-  WBXML.CompileCodepages(codepages);
-
-  return codepages;
-}));
